update backend (AddPlayer api)

+++ ttf-backend/src/main/java/io/aleksb/springjwt/ttfmain/repository/LeagueRepository.java
@@ -3,5 +3,5 @@ package io.aleksb.springjwt.ttfmain.repository;
 import io.aleksb.springjwt.ttfmain.models.League;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeagueRepository extends JpaRepository<League, Long> {
+public interface LeagueRepository extends JpaRepository<League, Integer> {
 }
